power draw issue

+++ include/status.h
@@ -19,7 +19,7 @@ extern int8_t really_connected;
 extern char ip_address[21];
 extern gt_mode_t current_mode;
 extern char last_error[50];
-extern uint16_t scan_counter;
+extern uint32_t scan_counter;
 
 void set_mode_tap_out();
 void set_mode_tap_in();
